add tests for empty frames
