+++ tests/empty_frames.csv
@@ -1,0 +1,2 @@
+﻿image_path,xmin,ymin,xmax,ymax,label
+JetportSouth_03_08_2021_382.png,0,0,0,0,Bird
